Added ExitedEvents for links whose sockets fail on creation.
When we try to link to a remote process, we will send the linkee an
`ExitedEvent` when the link is broken or if the connection fails.
This patch adds an `ExitedEvent` when the socket creation step fails.
This is logically equivalent to having the connection step fail.

Because this is an entirely unexpected case, the log level should be
a WARNING or higher.

Review: https://reviews.apache.org/r/52180/