Prevented a race when relinking with SSL downgrade enabled.
If there are two or more actors, on the same OS process, that try to
relink at the same time, there is a potential CHECK failure when SSL
downgrade is also enabled.  The following interleaving is problematic:
|             Actor A             |             Actor B             |
|---------------------------------+---------------------------------|
|        Starts to relink.        |        Starts to relink.        |
|        Creates socket A.        |                                 |
|   Replaces link with Socket A.  |                                 |
| Tries to connect with Socket A. |                                 |
|                                 |        Creates socket B.        |
|  Connection fails due to SSL.   |   Replaces link with Socket B.  |
|     Attempts to downgrade.      |                                 |
|     Tries to replace link.      |                                 |
The last step in the interleaving fails because we assert that the
socket we are swapping out exists in the `SocketManager`s state.

Review: https://reviews.apache.org/r/52182/