Optimized and simplified `DRFSorter::allocated()`.
When an allocation is made to a client, we want to update the share and
the count of allocations made to that client. This was previously done
separately, resulting in two std::set lookups, removals, and
reinsertions. By improving `DRFSorter::updateShare` to support
incrementing the number of client-allocations, we can do both updates
with a single sequence of std::set lookup/remove/insert operations.

Review: https://reviews.apache.org/r/56352