Updated mesos containerizer to ignore GPU isolator creation failure.
This cherry-pick will be maintained in DC/OS to avoid the problem of
requiring NVML to be installed on *every* node in the cluster, even if
GPUs aren't available on them.

Without this patch, we would need some sort of per-agent configuration
to only enable the GPU isolator on agents that had NVML intstalled on
them, and that is currently difficult (or undesirable) to do in DC/OS
today.

This is a decent compromise.