Added '--filter_gpu_resources' flag to the mesos master.
When set to true, this flag will cause the mesos master to filter
offers containing GPU resources by only sending them to frameworks
that opt into the 'GPU_RESOURCES' framework capability. When set to
false, this flag will cause the master to not filter offers containing
GPU resources, and indiscriminately send them to all frameworks
whether they set the 'GPU_RESOURCES' capability or not.  This flag is
meant as a temporary workaround towards the eventual deprecation of
the 'GPU_RESOURCES' capability.

Please see the following link for more information:
https://www.mail-archive.com/dev@mesos.apache.org/msg37571.html